src/xml/transform.ts
import { XMLDataBindingState } from '../../spec/XMLDataBinding';
import { CustomXMLPartState } from '../../spec/CustomXMLParts';
import { FormulaError } from '../errors';

export function transformXMLContent(
  content: string,
  stylesheet: string
): string {
  try {
    // TODO: Add XSLT transformation when available
    // For now, return original content
    return content;
  } catch (error) {
    throw new FormulaError('Error transforming XML content', error);
  }
}

export function mapXMLData(
  state: XMLDataBindingState,
  content: string,
  mapping: Map<string, string>
): Record<string, any> {
  try {
    const result: Record<string, any> = {};
    const parser = new DOMParser();
    const xmlDoc = parser.parseFromString(content, 'text/xml');

    for (const [xpath, target] of mapping.entries()) {
      const nodes = xmlDoc.evaluate(
        xpath,
        xmlDoc,
        null,
        XPathResult.ANY_TYPE,
        null
      );
      
      let node = nodes.iterateNext();
      if (node) {
        result[target] = node.textContent;
      }
    }

    return result;
  } catch (error) {
    throw new FormulaError('Error mapping XML data', error);
  }
}

export function generateXMLSchema(
  state: CustomXMLPartState,
  content: string
): string {
  try {
    // TODO: Add schema generation when available
    // For now, return a basic schema
    return `<?xml version="1.0" encoding="UTF-8"?>
<xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema">
  <!-- Generated schema placeholder -->
</xs:schema>`;
  } catch (error) {
    throw new FormulaError('Error generating XML schema', error);
  }
}

export function applyXMLTransformation(
  state: CustomXMLPartState,
  content: string,
  transformId: string
): string {
  try {
    const transform = state.operations.transform.stylesheets.get(transformId);
    if (!transform) {
      throw new Error(`Transform '${transformId}' not found`);
    }

    // TODO: Add XSLT transformation when available
    // For now, return original content
    return content;
  } catch (error) {
    throw new FormulaError('Error applying XML transformation', error);
  }
}